requirements.txt
Django==5.0.8
djangorestframework==3.15.2
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.1.1
django-cors-headers==4.4.0
python-decouple==3.8
pandas==2.2.3